adapted to cobra
