+++ cmd/retry.go
@@ -5,36 +5,64 @@ Copyright © 2022 NAME HERE <EMAIL ADDRESS>
 package cmd
 
 import (
-	"fmt"
+	"os"
+	"path/filepath"
 
 	"github.com/spf13/cobra"
+	"github.com/yohamta/dagu/agent"
+	"github.com/yohamta/dagu/internal/config"
+	"github.com/yohamta/dagu/internal/database"
+	"github.com/yohamta/dagu/internal/utils"
 )
+
+var req = ""
 
 // retryCmd represents the retry command
 var retryCmd = &cobra.Command{
 	Use:   "retry",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("retry called")
+	Short: "dagu retry --req=<request-id> <config>",
+	RunE: func(cmd *cobra.Command, args []string) error {
+		f, _ := filepath.Abs(args[0])
+		requestId := req
+		return retry(f, requestId)
 	},
 }
 
 func init() {
+	retryCmd.Flags().StringVar(&req, "req", "", "request-id")
+	retryCmd.MarkFlagRequired("req")
 	rootCmd.AddCommand(retryCmd)
+}
 
-	// Here you will define your flags and configuration settings.
+func retry(f, requestId string) error {
+	cl := &config.Loader{
+		HomeDir: utils.MustGetUserHomeDir(),
+	}
 
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// retryCmd.PersistentFlags().String("foo", "", "A help for foo")
+	db := database.New(database.DefaultConfig())
+	status, err := db.FindByRequestId(f, requestId)
+	if err != nil {
+		return err
+	}
 
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// retryCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	cfg, err := cl.Load(f, status.Status.Params)
+	if err != nil {
+		return err
+	}
+
+	a := &agent.Agent{
+		AgentConfig: &agent.AgentConfig{
+			DAG: cfg,
+			Dry: false,
+		},
+		RetryConfig: &agent.RetryConfig{
+			Status: status.Status,
+		},
+	}
+
+	listenSignals(func(sig os.Signal) {
+		a.Signal(sig)
+	})
+
+	return a.Run()
 }
