+++ cmd/retry_tmp.go
@@ -1,13 +1,7 @@
 package cmd
 
 import (
-	"os"
 	"path/filepath"
-
-	"github.com/yohamta/dagu/agent"
-	"github.com/yohamta/dagu/internal/config"
-	"github.com/yohamta/dagu/internal/database"
-	"github.com/yohamta/dagu/internal/utils"
 
 	"github.com/urfave/cli/v2"
 )
@@ -30,37 +24,4 @@ func newRetryCommand() *cli.Command {
 			return retry(f, requestId)
 		},
 	}
-}
-
-func retry(f, requestId string) error {
-	cl := &config.Loader{
-		HomeDir: utils.MustGetUserHomeDir(),
-	}
-
-	db := database.New(database.DefaultConfig())
-	status, err := db.FindByRequestId(f, requestId)
-	if err != nil {
-		return err
-	}
-
-	cfg, err := cl.Load(f, status.Status.Params)
-	if err != nil {
-		return err
-	}
-
-	a := &agent.Agent{
-		AgentConfig: &agent.AgentConfig{
-			DAG: cfg,
-			Dry: false,
-		},
-		RetryConfig: &agent.RetryConfig{
-			Status: status.Status,
-		},
-	}
-
-	listenSignals(func(sig os.Signal) {
-		a.Signal(sig)
-	})
-
-	return a.Run()
 }
