+++ cmd/start.go
@@ -5,36 +5,47 @@ Copyright © 2022 NAME HERE <EMAIL ADDRESS>
 package cmd
 
 import (
-	"fmt"
+	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/yohamta/dagu/agent"
+	"github.com/yohamta/dagu/internal/config"
+	"github.com/yohamta/dagu/internal/utils"
 )
+
+var paramsStart = ""
 
 // startCmd represents the start command
 var startCmd = &cobra.Command{
 	Use:   "start",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("start called")
+	Short: "dagu start [--params=\"<params>\"] <config>",
+	RunE: func(cmd *cobra.Command, args []string) error {
+		cl := &config.Loader{
+			HomeDir: utils.MustGetUserHomeDir(),
+		}
+		config_file_path := args[0]
+		cfg, err := cl.Load(config_file_path, paramsStart)
+		if err != nil {
+			return err
+		}
+		return start(cfg)
 	},
 }
 
 func init() {
+	startCmd.Flags().StringVar(&params, "params", "", "parameters")
 	rootCmd.AddCommand(startCmd)
+}
 
-	// Here you will define your flags and configuration settings.
+func start(cfg *config.Config) error {
+	a := &agent.Agent{AgentConfig: &agent.AgentConfig{
+		DAG: cfg,
+		Dry: false,
+	}}
 
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// startCmd.PersistentFlags().String("foo", "", "A help for foo")
+	listenSignals(func(sig os.Signal) {
+		a.Signal(sig)
+	})
 
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// startCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	return a.Run()
 }
