+++ cmd/status_tmp.go
@@ -1,11 +1,7 @@
 package cmd
 
 import (
-	"log"
-
 	"github.com/yohamta/dagu/internal/config"
-	"github.com/yohamta/dagu/internal/controller"
-	"github.com/yohamta/dagu/internal/models"
 	"github.com/yohamta/dagu/internal/utils"
 
 	"github.com/urfave/cli/v2"
@@ -27,16 +23,4 @@ func newStatusCommand() *cli.Command {
 			return queryStatus(cfg)
 		},
 	}
-}
-
-func queryStatus(cfg *config.Config) error {
-	status, err := controller.New(cfg).GetStatus()
-	if err != nil {
-		return err
-	}
-	res := &models.StatusResponse{
-		Status: status,
-	}
-	log.Printf("Pid=%d Status=%s", res.Status.Pid, res.Status.Status)
-	return nil
 }
