+++ cmd/stop_tmp.go
@@ -1,11 +1,8 @@
 package cmd
 
 import (
-	"log"
-
 	"github.com/urfave/cli/v2"
 	"github.com/yohamta/dagu/internal/config"
-	"github.com/yohamta/dagu/internal/controller"
 	"github.com/yohamta/dagu/internal/utils"
 )
 
@@ -25,10 +22,4 @@ func newStopCommand() *cli.Command {
 			return stop(cfg)
 		},
 	}
-}
-
-func stop(cfg *config.Config) error {
-	c := controller.New(cfg)
-	log.Printf("Stopping...")
-	return c.Stop()
 }
