+++ cmd/version.go
@@ -8,33 +8,22 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
+	"github.com/yohamta/dagu/internal/constants"
 )
 
 // versionCmd represents the version command
 var versionCmd = &cobra.Command{
 	Use:   "version",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("version called")
+	Short: "dagu version",
+	RunE: func(cmd *cobra.Command, args []string) error {
+		if constants.Version != "" {
+			fmt.Println(constants.Version)
+			return nil
+		}
+		return fmt.Errorf("failed to get version")
 	},
 }
 
 func init() {
 	rootCmd.AddCommand(versionCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// versionCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// versionCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
